Add Swagger and add put for teams and player crud

+++ src/teams/teams.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateTeamDto } from './dto/create-team.dto';
+import { UpdateTeamDto } from './dto/update-team.dto';
 
 @Injectable()
 export class TeamsService {
@@ -20,14 +21,20 @@ export class TeamsService {
     return team;
   }
 
-  async update(id: number, updateData: Partial<CreateTeamDto>) {
+  async update(id: number, data: UpdateTeamDto) {
+    const team = await this.prisma.team.findUnique({ where: { id } });
+    if (!team) throw new NotFoundException('Time não encontrado');
+
     return this.prisma.team.update({
       where: { id },
-      data: updateData,
+      data,
     });
   }
 
   async remove(id: number) {
+    const team = await this.prisma.team.findUnique({ where: { id } });
+    if (!team) throw new NotFoundException('Time não encontrado');
+
     return this.prisma.team.delete({ where: { id } });
   }
 }
